refactor branches2
