+++ src/modules/jinro/classes/Player.ts
